% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANN_parameters.R
\docType{data}
\name{ANN_parameters}
\alias{ANN_parameters}
\title{Parameters for Brey's ANN model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 23 columns.
}
\usage{
ANN_parameters
}
\description{
Conversion factors and categorical variables for Brey's ANN model.
\itemize{
\item \code{Group} The graphical.
\item \code{Taxon} Taxa in the recorded form.
\item \code{ConFac_j2mgwm} The conversion factors from mg WM to Joule.
\item \code{ConFac_j2mgc} The conversion factors from mg C to Joule.
\item \code{Taxonomic categories} Mollusca, Annelida, Crustacea, Echinodermata, Insecta.
\item \code{Lifestyle categories} Infauna, Sessile, Crawler, FacultativeSwimmer.
\item \code{Diet categories} Herbivore, Omnivore, Carnivore.
\item \code{Habitat categories} Lake, River, Marine, Subtidal.
\item \code{Exploited} The presence of exploitation.
\item \code{Comment} Comments on the decisions.
\item \code{Reference} Reference of the comments.
}
}
\keyword{datasets}
