% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_biovolume.R
\name{calculate_biovolume}
\alias{calculate_biovolume}
\title{calculate biovolume of each individual}
\usage{
calculate_biovolume(data)
}
\arguments{
\item{data}{A long format data with Length (L), Width (W), and Method (Method) as their variables.}
}
\value{
The same data frame except that a new column called "Size" contains the volume data.
}
\description{
Calculates biovolume of the individuals with its assigned methods.
}
\details{
Note that this function will first remove all the entris in the \code{Size} column.
Another thing is that this function is not flexible enough. Further revision is needed.
}
\examples{
a <- data.frame(
  L = c(10, 5, 6, 78, 1, 6),
  W = c(0.1, 3, 4, 1, 2, 3),
  Method = c("LWR", "LWR", "Cylinder", "Cylinder", "Cone", "Elliposid"),
  C = c(0.53, 0.45, NA, NA, NA, NA)
)
calculate_biovolume(a)
}
