% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cone.R
\name{Cone}
\alias{Cone}
\title{Estimate biovolume with cone equation}
\usage{
Cone(Length, Width)
}
\arguments{
\item{Length}{Length of an organism}

\item{Width}{Width of an organism}
}
\value{

}
\description{
Estimate the biovolume of individuals with cone equation.
}
\examples{
Cone(30, 1)
}
