% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_factors.R
\docType{data}
\name{conversion_factors}
\alias{conversion_factors}
\title{Conversion factors for Brey's ANN model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 108 rows and 10 columns.
}
\usage{
conversion_factors
}
\description{
Conversion factors derived from conversion factors.
\itemize{
\item \code{Group} The graphical.
\item \code{Taxon} Taxa in the recorded form.
\item \code{Conversion_factors} The type of conversion factors.
\item \code{N} Number of data.
\item \code{Mean} The mean value of conversion factors.
\item \code{S.D.} The standard deviation of the conversion factors.
\item \code{Median} The median of the conversion factors.
\item \code{Accquired_taxonomic_level} The taxonomic level of where the conversion factors were obtained.
\item \code{Additional_diet} The additional diet feature if the taxonomic level was set to the highest level.
\item \code{Comment} Comment about the whole row.
}
}
\keyword{datasets}
