% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cylinder.R
\name{Cylinder}
\alias{Cylinder}
\title{Estimate biovolume with cylinder equation}
\usage{
Cylinder(Length, Width)
}
\arguments{
\item{Length}{Length of an organism}

\item{Width}{Width of an organism}
}
\value{
The volume
}
\description{
Estimate the biovolume of individuals with cylinder equation.
}
\examples{
Cylinder(15, 0.75)
}
