% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ellipsoid.R
\name{Ellipsoid}
\alias{Ellipsoid}
\title{Estimate biovolume with ellipsoid equation}
\usage{
Ellipsoid(Length, Width)
}
\arguments{
\item{Length}{Length of an organism}

\item{Width}{Width of an organism}
}
\value{

}
\description{
Estimate the biovolume of individuals with ellipsoid equation.
}
\examples{
Ellipsoid(10, 5)
}
