% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANN_parameters.R
\docType{data}
\name{ANN_parameters}
\alias{ANN_parameters}
\title{ANN parameters for Brey's 2010 model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 23 columns.
}
\usage{
ANN_parameters
}
\description{
ANN parameters for Brey's 2010 model
}
\details{
The condition and unit of each column were listed below.
\itemize{
\item{Group} Grouped taxon for simple visualization
\item{Taxon} Recorded grouping taxa of the specimen
\item{ConFac_j2mgwm} The mean value of mgWM-to-J Conversion factors collected from Brey et al., 2010.
\item{ConFac_j2mgc} The mean value of mgWM-to-J Conversion factors collected from Brey et al., 2010.
\item{Mollusca} Taxonomic categorical varaibles.
\item{Annelida} Taxonomic categorical varaibles.
\item{Crustacea} Taxonomic categorical varaibles.
\item{Echinodermata} Taxonomic categorical varaibles.
\item{Insecta} Taxonomic categorical varaibles.
\item{Infauna} Lifestyle categorical varaibles.
\item{Sessile} Lifestyle categorical varaibles.
\item{Crawler} Lifestyle categorical varaibles.
\item{FacultativeSwimmer} Lifestyle categorical varaibles.
\item{Herbivore} Diet categorical varaibles.
\item{Omnivore} Diet categorical varaibles.
\item{Carnivore} Diet categorical varaibles.
\item{Lake} Habitat categorical varaibles.
\item{River} Habitat categorical varaibles.
\item{Marine} Habitat categorical varaibles.
\item{Subtidal} Habitat categorical varaibles.
\item{Exploited} Categorical varaible for exploitation.
\item{Comment} Personal comments.
\item{Reference}
}
}
\author{
Yen-Ting Chen \email{r08241220@ntu.edu.tw}
}
\keyword{datasets}
